REMOVE POST TYPE FOR REUSABLE BLOCKS FROM CLEAR THEME CUSTOMISATIONS
The clear-theme-customisations.sh script clears down posts associated
with four post types, wp_block, wp_template, wp_template_part and
wp_global_styles. The wp_template, wp_template_part and wp_global posts
are automatically restored to the latest versions of the files contained
with the theme when they are cleared from the database. The same is not
true of posts of the wp_block post type, which must be manually (as
things stand) re-imported. Futhermore when resuable blocks are used they
are referenced by their ID, which will change if they are deleted and
re-imported. So, best not to mass delete reusable block posts in a
script that is really about clearing theme customisations and should not
be doing this also.